Version 1.1.3
Simple Skyblock now handles world generation!
Simple Skyblock can now run without any dependencies! It now handles world generation, biome setting and all that fun stuff.
First of all, if you want to have the default world be your Skyblock world, you have to add the following to your bukkit.yml file:
Code (Text):
worlds:
  world:
    generator: SimpleSkyblock
It basically just handles generaion for 'world' over to the plugin so it can do it's magic.
If you have another world specified for your Skyblock world, it will automatically generate it, and teleport users to it when they want to visit their island. There is no need to add something like the above to your bukkit.yml file.
There is a new item in the config.yml file, called "BIOME". Basically, it is the biome that the whole blank world will be when it generates. Although you can change this after the world has been generated, it will only apply to newly generated chunks.

Special thanks to robertlit for helping with the world generation.

As always, if you find any bugs or have a feature request, reach out to me in some way and I will get to it as soon as I can.