Version 1.1.5
Fixed a massive bug, my bad
The garbage collector was garbage because it had a typo. Sorry about that.

As always, if you ever find a bug, or have a new feature suggestion, feel free to send me a message and I'll get to it as soon as I can
- garet90

Added garbage collector
To prevent overflowing of active users, when there are 25% more profiles open for reading/writing than users that are online, they are pushed out of memory (they can be opened again if they are requested). Player data mostly get opened when trying to do /is visit commands and have no use staying open if they are using a substantial amount of memory.
Also, I found some typos in the dialogue that were fixed, and some bugs were fixed.

As always, if you find any bugs or have any feature requests, shoot me a message!
- garet90