Version 1.2.1
New config, special mechanics, bug fixes and more!
New Config
The config file is looking better than ever! It is now more organized, and provides a description for each of the settings. New settings will no longer be posted in the updates unless they are major, and can be found with an explanation in config.yml, or on the main page of this plugin under config.yml.

New mechanics
New optional mechanics have been added to help players obtain objects that would otherwise be unobtainable. These can be disabled, and are found in config.yml

Various bug fixes
Many bugs have been fixed related to island generation and economy. Most of them are undocumented, but you can see the file changes on GitHub.

As always, be sure to contact if you find any bugs or have any suggestions.