Version 1.2.2
Plugin has been rewritten!
Because of the sheer size of the plugin, SimpleSkyblock has been rewritten in a more object oriented design to aid in development. Because of this, much of the data structure and backbone of SimpleSkyblock has changed. It is highly recommended that you completely reset SimpleSkyblock by deleting it's data files and world when updating, but not mandatory.

Welcome to version 1.2.2! Although much of the plugin has changed, as it has been rewritten, you may not notice much of a difference other than performance, as almost all of the features are the same.

Timeouts on requests:
Island invites and trade requests now expire after 60 seconds, and the new data structure allows multiple to be sent to one player.

Bug fixes:
Many undocumented bugs have been fixed.

Plugin compatibility:
Plugins that interact with SimpleSkyblock are easier to write now, a wiki will be coming soon on the GitHub page.